+++ Tokens/solana-dvp/contracts/src/test/kotlin/net/corda/samples/solanadvp/contracts/ContractTests.kt
@@ -1,43 +1,57 @@
 package net.corda.samples.solanadvp.contracts
 
 import com.r3.corda.lib.tokens.contracts.commands.Create
-import net.corda.core.contracts.Amount
 import net.corda.core.contracts.UniqueIdentifier
 import net.corda.core.identity.CordaX500Name
-import net.corda.samples.solanadvp.states.DeliveryState
+import net.corda.samples.solanadvp.states.StockState
 import net.corda.testing.core.TestIdentity
 import net.corda.testing.node.MockServices
 import net.corda.testing.node.ledger
 import org.junit.Test
+import java.math.BigDecimal
+import java.util.Date
 
 class ContractTests {
     private val ledgerServices = MockServices()
-    val operator = TestIdentity(CordaX500Name(organisation = "Alice", locality = "TestLand", country = "US"))
+    val issuer = TestIdentity(CordaX500Name(organisation = "Alice", locality = "TestLand", country = "US"))
 
     //sample tests
     @Test
     fun `Price must be greater than zero`() {
-        val tokenPass = DeliveryState(
-            UniqueIdentifier(),
-            listOf(operator.party),
-            Amount.parseCurrency("1000 USD")
+
+        val tokenPass = StockState(
+            issuer.identity.party,
+            "AAPL",
+            "Apple",
+            "USD",
+            BigDecimal(273.12),
+            BigDecimal.ZERO,
+            Date(),
+            Date(),
+            UniqueIdentifier()
         )
-        val tokenFail = DeliveryState(
-            UniqueIdentifier(),
-            listOf(operator.party),
-            Amount.parseCurrency("0 USD")
+        val tokenFail = StockState(
+            issuer.identity.party,
+            "AAPL",
+            "Apple",
+            "USD",
+            BigDecimal.ZERO,
+            BigDecimal.ZERO,
+            Date(),
+            Date(),
+            UniqueIdentifier()
         )
         ledgerServices.ledger {
             transaction {
-                output(DeliveryContract.CONTRACT_ID, tokenFail)
-                command(operator.publicKey, Create())
+                output(StockContract.CONTRACT_ID, tokenFail)
+                command(issuer.publicKey, Create())
                 this.fails()
             }
         }
         ledgerServices.ledger {
             transaction {
-                output(DeliveryContract.CONTRACT_ID, tokenPass)
-                command(operator.publicKey, Create())
+                output(StockContract.CONTRACT_ID, tokenPass)
+                command(issuer.publicKey, Create())
                 this.verifies()
             }
         }
