+++ Tokens/solana-dvp/workflows/build.gradle
@@ -71,6 +71,21 @@ dependencies {
     cordaProvided "$tokens_release_group:tokens-workflows:$tokens_release_version"
 }
 
+
+tasks.withType(Test).configureEach { t ->
+    systemProperty 'co.paralleluniverse.fibers.verifyInstrumentation', 'false'
+    systemProperty 'co.paralleluniverse.fibers.detectRunawayFibers', 'false'
+
+    // Remove existing quasar -javaagent if something (e.g. corda/quasar-utils) already adds it
+    def currentArgs = (t.jvmArgs ?: [])
+    t.jvmArgs = currentArgs.findAll { !(it.toString().contains('quasar') && it.toString().contains('-javaagent:')) }
+
+    // Add Quasar agent with allowMonitors (=m) to fix kotlin/SynchronizedLazyImpl issues
+    def quasarJar = configurations.quasar.singleFile
+    t.jvmArgs "-javaagent:${quasarJar}=m"
+}
+
+
 task integrationTest(type: Test, dependsOn: []) {
     useJUnitPlatform()
     testClassesDirs = sourceSets.integrationTest.output.classesDirs
